moving Java to a single file
