--- java/src/bike_europe/Coordinate.java
@@ -1,7 +1,0 @@
-package bike_europe;
-
-public interface Coordinate {
-  public Double getLat();
-  public Double getLng();
-  public Double kilometersTo(Coordinate other);
-}
